build service tag in advance; move filter methods to state retriever; rename FilteredState method to LoadBalancerState
