+++ anx/provider/loadbalancer/reconciliation/reconcile_binds.go
@@ -1,9 +1,6 @@
 package reconciliation
 
 import (
-	"fmt"
-	"sort"
-
 	"go.anx.io/go-anxcloud/pkg/api/types"
 	"go.anx.io/go-anxcloud/pkg/utils/object/compare"
 
@@ -11,34 +8,6 @@ import (
 )
 
 const bindResourceTypeIdentifier = "bd24def982aa478fb3352cb5f49aab47"
-
-func (r *stateRetrieverImpl) storePublicAddress(lbID, addr string) {
-	if idx := sort.SearchStrings(r.loadBalancers[lbID].state.publicAddresses, addr); idx >= len(r.loadBalancers[lbID].state.publicAddresses) || r.loadBalancers[lbID].state.publicAddresses[idx] != addr {
-		r.loadBalancers[lbID].state.publicAddresses = append(r.loadBalancers[lbID].state.publicAddresses, addr)
-		sort.Strings(r.loadBalancers[lbID].state.publicAddresses)
-	}
-}
-
-func (r *stateRetrieverImpl) filterBinds(lbID string, allBinds []*lbaasv1.Bind) ([]*lbaasv1.Bind, error) {
-	ret := make([]*lbaasv1.Bind, 0, len(allBinds))
-
-	// Binds and Servers are filtered for our LoadBalancer here, after we hopefully retrieved their Frontends and Backends already
-	for _, bind := range allBinds {
-		idx, err := compare.Search(lbaasv1.Frontend{Identifier: bind.Frontend.Identifier}, r.loadBalancers[lbID].state.frontends, "Identifier")
-		if err != nil {
-			return nil, fmt.Errorf("error checking if Binds belongs to one of our frontends: %w", err)
-		} else if idx != -1 {
-			ret = append(ret, bind)
-			r.sortObjectIntoStateArray(lbID, bind)
-
-			if bind.Address != "" {
-				r.storePublicAddress(lbID, bind.Address)
-			}
-		}
-	}
-
-	return ret, nil
-}
 
 func (r *reconciliation) reconcileBinds() (toCreate, toDestroy []types.Object, err error) {
 	targetBinds := make([]*lbaasv1.Bind, 0, len(r.externalAddresses)*len(r.ports))
