+++ anx/provider/loadbalancer/reconciliation/reconcile_servers.go
@@ -1,8 +1,6 @@
 package reconciliation
 
 import (
-	"fmt"
-
 	"go.anx.io/go-anxcloud/pkg/api/types"
 	"go.anx.io/go-anxcloud/pkg/utils/object/compare"
 
@@ -10,22 +8,6 @@ import (
 )
 
 const serverResourceTypeIdentifier = "01f321a4875446409d7d8469503a905f"
-
-func (r *stateRetrieverImpl) filterServers(lbID string, allServers []*lbaasv1.Server) ([]*lbaasv1.Server, error) {
-	ret := make([]*lbaasv1.Server, 0, len(allServers))
-
-	for _, server := range allServers {
-		idx, err := compare.Search(lbaasv1.Backend{Identifier: server.Backend.Identifier}, r.loadBalancers[lbID].state.backends, "Identifier")
-		if err != nil {
-			return nil, fmt.Errorf("error checking if Server belongs to one of our frontends: %w", err)
-		} else if idx != -1 {
-			ret = append(ret, server)
-			r.sortObjectIntoStateArray(lbID, server)
-		}
-	}
-
-	return ret, nil
-}
 
 func (r *reconciliation) reconcileServers() (toCreate, toDestroy []types.Object, err error) {
 	targetServers := make([]*lbaasv1.Server, 0, len(r.ports)*len(r.targetServers))
